+++ src/World/League.java
@@ -2,9 +2,9 @@ package World;
 
 import java.io.File;
 import java.io.FileNotFoundException;
-import java.lang.reflect.Array;
 import java.util.*;
 import Comparator.CompareTeamWinPct;
+import World.Player.PlayerQB;
 
 public class League {
     public static Team[][][] leagueList; //0-1 World.Conference | 0-3 World.Division | 0-3 World.Team
