+++ src/World/Player/PlayerILB.java
@@ -1,4 +1,6 @@
-package World;
+package World.Player;
+import World.Team;
+
 import java.util.Random;
 
 public class PlayerILB extends Player{
