+++ src/World/Player/PlayerOG.java
@@ -1,8 +1,11 @@
-package World;
+package World.Player;
+
+import World.Team;
+
 import java.util.Random;
 
-public class PlayerOC extends Player{
-  private static final String [] playerTypes = {"Mauler", "Agile"};
+public class PlayerOG extends Player{
+  private static final String [] playerTypes = {"Mauler", "Agile", "Puller"};
   private static final int numTypes = playerTypes.length;
 
 
@@ -14,31 +17,64 @@ public class PlayerOC extends Player{
 
 
   //Use this constructor for populating teams during league creation
-  public PlayerOC(Team t){
+  public PlayerOG(Team t){
 
-    whiteWeight = 65;
-    blackWeight = 35;
+    whiteWeight = 47;
+    blackWeight = 53;
 
     this.race = findRace();
     getNames();
 
 
     team = t;
-    position = "Offensive Center";
+    position = "Offensive Guard";
     this.playerType = playerTypes[(int) (Math.random() * numTypes)];
 
     switch (this.playerType){
       case "Mauler" -> getMaulerStats();
       case "Agile" -> getAgileStats();
+      case "Puller" -> getPullerStats();
     }
     this.overall = getOverall();
 
 
   }
 
+  private void getPullerStats() {
+    Random rand = new Random();
+    int maxHeight = 76;
+    int minHeight = 74;
+    this.heightIn = rand.nextInt(minHeight, maxHeight);
+
+    int maxWeight = 325;
+    int minWeight = 298;
+    this.weight = rand.nextInt(minWeight, maxWeight);
+
+    int maxSpeed = 70;
+    int minSpeed = 62;
+    this.speed = rand.nextInt(minSpeed, maxSpeed);
+
+    int maxPassBlock = 83;
+    int minPassBlock = 74;
+    this.passBlock = rand.nextInt(minPassBlock, maxPassBlock);
+
+    int maxRunBlock = 93;
+    int minRunBlock = 82;
+    this.runBlock = rand.nextInt(minRunBlock, maxRunBlock);
+
+    int maxQuickness = 69;
+    int minQuickness = 50;
+    this.quickness = rand.nextInt(minQuickness, maxQuickness);
+
+    int maxStrength = 97;
+    int minStrength = 88;
+    this.strength = rand.nextInt(minStrength, maxStrength);
+
+  }
+
   private int getOverall() {
-    int overall = (int) ( + passBlockWeight*passBlock + runBlockWeight*runBlock + strength*strengthWeight + quickness*quicknessWeight);
-    return overall;
+    int o = (int) ( + passBlockWeight*passBlock + runBlockWeight*runBlock + strength*strengthWeight + quickness*quicknessWeight);
+    return o;
   }
 
   private void getMaulerStats() {
